+++ src/main/java/com/bush/pharmacy_web_app/controllers/http/admin/AdminOrderController.java
@@ -1,4 +1,4 @@
-package com.bush.pharmacy_web_app.controllers.admin;
+package com.bush.pharmacy_web_app.controllers.http.admin;
 
 import com.bush.pharmacy_web_app.repository.entity.order.OrderStatus;
 import com.bush.pharmacy_web_app.service.OrderService;
