+++ src/main/resources/static/css/admin/order.css
@@ -128,7 +128,7 @@ body {
     margin-top: 1rem;
 }
 
-.fluent-button:hover {
+.order-more-det:hover {
     background: #3650b5;
 }
