+++ src/main/resources/templates/admin/order-management.html
@@ -8,7 +8,7 @@
 <body>
     <div th:replace="~{/admin/sidebar/sidebar.html :: sidebar}"></div>
     <main class="main-content">
-        <a class="back-link" th:href="@{/admin/orders}">Назад к списку продуктов</a>
+        <a class="back-link" th:href="@{/admin/orders}">Назад к списку заказов</a>
         <div>
             <div class="order-header">
                 <h1>Заказ - [(${order.id})]</h1>
