+++ src/main/resources/templates/admin/order.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>Управление заказами</title>
     <link href="/css/admin/order.css" rel="stylesheet" type="text/css">
+    <script src="/js/admin/order.js" rel="script" type="text/javascript"></script>
 </head>
 <body>
     <div th:replace="~{/admin/sidebar/sidebar.html :: sidebar}"></div>
@@ -25,7 +26,7 @@
                         <span>Итого:</span>
                         <p>[(${order.result})] ₽</p>
                     </div>
-                    <button class="order-more-det">Подробнее</button>
+                    <button class="order-more-det" id="order-more-det" th:data-id="${order.id}">Подробнее</button>
                 </div>
             </div>
         </div>
